move utilities around
